updated
git-svn-id: file:///home/roberto/Git/svn@11541 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: c06bd835acb78d43aa9354f645dd595d3bd53ec7

+++ trunk/xmipp/libraries/reconstruction/ctf_group.cpp
@@ -572,7 +572,7 @@ void ProgCtfGroup::writeOutputToDisc()
     ctfInfo.aggregate(sortedCtfMD,aggregateOperations,operateLabels,resultLabels);
     ctfInfo.setComment("N. of micrographs, N. of particles, min defocus, max defocus and avg defocus");
     ctfInfo.write("groups@"+fn_root+"Info.xmd");
-    int numberDefGroups=ctfInfo.size();
+    size_t numberDefGroups=ctfInfo.size();
     MetaData MD;
     size_t idctf = MD.addObject();
     MD.setValue(MDL_COUNT,numberDefGroups,idctf);
